+++ CsprojToAsmdef/Assets/Sample/Hello.cs
@@ -1,4 +1,6 @@
-﻿using TMPro;
+﻿using SampleUnityDependency;
+using SomeOtherDependency;
+using TMPro;
 using UnityEngine;
 
 namespace Sample
@@ -10,8 +12,8 @@ namespace Sample
 
         private void Start()
         {
-            Debug.Log("Hello");
-            _text.text = "Hello";
+            CustomLogger.LogWarning("My dude!");
+            _text.text = Messenger.GetMessage();
         }
     }
 }